app/requirements.txt
Celery==4.0.2
Django==1.11.3
DjangoRestFramework==3.6.3
django-cors-headers==2.1.0
psycopg2==2.7.1
